nginx, redis: fixed

+++ srcs/requirements/wordpress/conf/setup.sh
@@ -1,12 +1,15 @@
 #!/bin/bash
 
+set -e
+WP_PATH="/var/www/html"
+
 # Wait for MariaDb to be ready
 until mysql -h mariadb -u ${MYSQL_USER} -p${MYSQL_PASSWORD} -e "SELECT 1" &>/dev/null; do
-	echo "Waiting for MariaDb to be ready..."
+	echo "[WP] Waiting for MariaDb to be ready..."
 	sleep 3
 done
 
-cd /var/www/html
+cd $WP_PATH
 
 # Download WordPress if it doesn't exist
 if [ ! -f wp-config.php ]; then
@@ -43,13 +46,33 @@ if [ ! -f wp-config.php ]; then
 fi
 
 
-# Redis (safe & optional) (bonus)
-wp redis enable --allow-root --path=/var/www/html || \
-    echo "[WordPress] Redis not available, fallback to DB"
 
-# If Redis is not enabled, ensure no drop-in exists
-wp redis status --allow-root --path=/var/www/html | grep -q "Connected" || \
-    rm -f wp-content/object-cache.php
+# Redis bonus configuration
+echo "[WP] Waiting for Redis..."
+TIMEOUT=30
+while ! nc -z redis 6379 >/dev/null 2>&1; do
+    sleep 2
+    TIMEOUT=$((TIMEOUT-2))
+    if [ $TIMEOUT -le 0 ]; then
+        echo "[WP] Redis not available -> fallback to MariaDB"
+        rm -f wp-content/object-cache.php
+        exec php-fpm7.4 -F
+    fi
+done
+
+echo "[WP] Redis detected -> configuring"
+wp config set WP_REDIS_HOST redis --allow-root --path="$WP_PATH"
+wp config set WP_REDIS_PORT 6379 --raw --allow-root --path="$WP_PATH"
+wp config set WP_CACHE true --raw --allow-root --path="$WP_PATH"
+
+wp plugin install redis-cache \
+    --activate \
+    --allow-root --path="$WP_PATH" || true
+
+wp redis enable --allow-root --path="$WP_PATH" || true
+
+echo "[WP] Redis enabled successfully"
+
 
 
 # Start PHP-FPM in foreground
@@ -68,3 +91,10 @@ echo "Wordpress started successfully."
 # Redis is enabled optimistically.
 #  If it’s available, WordPress uses it.
 #  If it’s not, the object-cache drop-in is removed and WordPress falls back to MariaDB without interruption.
+#
+# - Wait MariaDB
+# - Install WP
+# - Wait Redis TCP 6379
+# - If Redis does not appear in 30s → continue without Redis
+# - If Redis appears → configure + enable
+# - Zero fatal errors
